tests/wpt/tests/speculation-rules/speculation-tags/prefetch-eagerness-pointer-down.https.html

<!DOCTYPE html>
<title>Sec-Speculation-Tags request headers</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="/common/utils.js"></script>
<script src="/html/browsers/browsing-the-web/remote-context-helper/resources/remote-context-helper.js"></script>
<script src="/speculation-rules/prerender/resources/utils.js"></script>
<script src="/speculation-rules/resources/utils.js"></script>
<script src="/speculation-rules/prefetch/resources/utils.sub.js"></script>
<script>
"use strict";

setup(() => assertSpeculationRulesIsSupported());

promise_test(async t => {
  const agent = await spawnWindow(t);
  const nextUrl = agent.getExecutorURL({ page: 2 });
  await agent.forceSpeculationRules({
      prefetch: [
        {source: "list", "tag": "conservative",
          "eagerness": "conservative", urls: [nextUrl]},
        {source: "list", "tag": "moderate",
          "eagerness": "moderate", urls: [nextUrl]}
      ]
  });

  const linkId = 'target';
  await agent.execute_script(async (linkId, nextUrl) => {
    add_link(linkId, nextUrl);
  }, [linkId, nextUrl]);

  await agent.execute_script(async (linkId) => {
    await start_non_immediate_prefetch_on_pointerdown(linkId);
  }, [linkId]);
  await agent.navigate(nextUrl);

  const headers = await agent.getRequestHeaders();
  assert_prefetched(headers, "must be prefetched");
  assert_equals(headers.sec_speculation_tags, "\"conservative\", \"moderate\"", "Sec-Speculation-Tags");
}, "Sec-Speculation-Tags prefetch with eagerness triggered by point down");

</script>
